requirements.txt
napari[all]
PyQt5
opencv-python
numpy
pandas
tifffile
imagecodecs
pyyaml
pyinstaller